Renamed ImportExportTab to WebSearchTab
