+++ src/main/java/org/jabref/gui/preferences/websearch/WebSearchTab.fxml
@@ -11,8 +11,8 @@
 <?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-         fx:controller="org.jabref.gui.preferences.importexport.ImportExportTab">
-    <Label styleClass="titleHeader" text="%Import and Export"/>
+         fx:controller="org.jabref.gui.preferences.websearch.WebSearchTab">
+    <Label styleClass="titleHeader" text="%Web search"/>
 
     <Label styleClass="sectionHeader" text="%General"/>
     <CheckBox fx:id="generateNewKeyOnImport" text="%Generate a new key for imported entries (overwriting their default)"/>
