+++ src/main/resources/l10n/JabRef_en.properties
@@ -2340,7 +2340,7 @@ Unprotect\ terms=Unprotect terms
 
 Generate\ a\ new\ key\ for\ imported\ entries\ (overwriting\ their\ default)=Generate a new key for imported entries (overwriting their default)
 Warn\ about\ duplicates\ on\ import=Warn about duplicates on import
-Import\ and\ Export=Import and Export
+
 Custom\ DOI\ URI=Custom DOI URI
 Use\ custom\ DOI\ base\ URI\ for\ article\ access=Use custom DOI base URI for article access
